Refactor engine subsystems into modules
