+++ engine/src/engine.cpp
@@ -1,7 +1,9 @@
-﻿// [UNCHANGED includes...]
-#include <engine/camera.hpp>
+﻿#include <engine/camera.hpp>
 #include <engine/config.hpp>
 #include <engine/engine.hpp>
+#include <engine/brick_manager.hpp>
+#include <engine/render_loop.hpp>
+#include <engine/resource_setup.hpp>
 #include <engine/gfx/memory.hpp>
 #include <engine/gfx/present_pipeline.hpp>
 #include <engine/gfx/ray_pipeline.hpp>
@@ -24,37 +26,16 @@
 #include <cmath>
 #include <cstdint>
 #include <filesystem>
-#include <fstream>
 #include <glm/gtc/matrix_inverse.hpp>
 #include <memory>
 #include <spdlog/spdlog.h>
-#include <thread>
 #include <vector>
-#include <unordered_map>
-
-#ifdef _WIN32
-#define NOMINMAX
-#include <windows.h>
-#endif
 
 namespace engine {
-
-static std::filesystem::path exe_dir() {
-#ifdef _WIN32
-  char buf[MAX_PATH]{};
-  DWORD n = GetModuleFileNameA(nullptr, buf, MAX_PATH);
-  return std::filesystem::path(std::string(buf, buf + n)).parent_path();
-#else
-  return std::filesystem::current_path();
-#endif
-}
 
 static constexpr uint32_t kStepsDown = 4;
 static constexpr float kRenderScale = 0.66f;
 
-static constexpr float kColorCompute[4]  = {0.0f, 0.6f, 1.0f, 1.0f};
-static constexpr float kColorGraphics[4] = {1.0f, 0.4f, 0.0f, 1.0f};
-static constexpr float kColorTransfer[4] = {0.4f, 1.0f, 0.4f, 1.0f};
 
 static float halton(uint32_t i, uint32_t b) {
   float f = 1.0f;
@@ -67,710 +48,8 @@ static float halton(uint32_t i, uint32_t b) {
   return r;
 }
 
-// Manager for streaming 3-D voxel bricks.  Each brick is a fixed size volume
-// stored inside a larger atlas texture.  Bricks are addressed by integer
-// coordinates in world space.  A separate lookup texture stores the mapping
-// from world-space brick coordinates to atlas indices.
-struct BrickManager {
-  struct BrickKey {
-    int x = 0, y = 0, z = 0;
-    bool operator==(const BrickKey &o) const {
-      return x == o.x && y == o.y && z == o.z;
-    }
-  };
-  struct BrickKeyHash {
-    size_t operator()(const BrickKey &k) const noexcept {
-      size_t h = std::hash<int>()(k.x);
-      h ^= std::hash<int>()(k.y) + 0x9e3779b9 + (h << 6) + (h >> 2);
-      h ^= std::hash<int>()(k.z) + 0x9e3779b9 + (h << 6) + (h >> 2);
-      return h;
-    }
-  };
-
-  struct Brick {
-    BrickKey key{};
-    uint32_t index = 0;
-    uint32_t last_used = 0;
-  };
-
-  VmaAllocator allocator = VK_NULL_HANDLE;
-  VkDevice device = VK_NULL_HANDLE;
-  uint32_t brick_size = 32;      // Edge length of a brick in voxels
-  uint32_t max_bricks = 64;      // Maximum number of bricks resident
-  glm::ivec3 map_dim{64};        // Dimensions of lookup texture in bricks
-
-  Image3D occ_atlas{};           // Atlas containing occupancy bricks
-  Image3D mat_atlas{};           // Atlas containing material bricks
-  Image3D index_img{};           // Lookup texture mapping coords->indices
-
-  UniqueImageView occ_view;            // Sampling view for occupancy atlas
-  UniqueImageView occ_storage_view;    // Storage view for occupancy atlas
-  UniqueImageView mat_view;            // Sampling view for material atlas
-  UniqueImageView mat_storage_view;    // Storage view for material atlas
-  UniqueImageView index_view;          // Sampling view for index texture
-  UniqueImageView index_storage_view;  // Storage view for index texture
-
-  std::unordered_map<BrickKey, Brick, BrickKeyHash> bricks;
-  std::vector<Brick> lru;              // Indexed by brick index
-  std::vector<uint32_t> free_list;     // Available atlas slots
-  std::vector<uint32_t> cpu_index;     // CPU copy of lookup texture
-
-  void init(VmaAllocator alloc, VkDevice dev, uint32_t bsize,
-            uint32_t maxb, glm::ivec3 map_dim_ = glm::ivec3(64)) {
-    allocator = alloc;
-    device = dev;
-    brick_size = bsize;
-    max_bricks = maxb;
-    map_dim = map_dim_;
-
-    occ_atlas = create_image3d(
-        allocator, brick_size, brick_size, brick_size * max_bricks,
-        VK_FORMAT_R8_UINT,
-        VK_IMAGE_USAGE_STORAGE_BIT | VK_IMAGE_USAGE_SAMPLED_BIT);
-    mat_atlas = create_image3d(
-        allocator, brick_size, brick_size, brick_size * max_bricks,
-        VK_FORMAT_R8_UINT,
-        VK_IMAGE_USAGE_STORAGE_BIT | VK_IMAGE_USAGE_SAMPLED_BIT);
-    index_img = create_image3d(
-        allocator, map_dim.x, map_dim.y, map_dim.z, VK_FORMAT_R32_UINT,
-        VK_IMAGE_USAGE_STORAGE_BIT | VK_IMAGE_USAGE_SAMPLED_BIT);
-
-    VkImageViewCreateInfo vi{VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO};
-    vi.viewType = VK_IMAGE_VIEW_TYPE_3D;
-    vi.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    vi.subresourceRange.levelCount = 1;
-    vi.subresourceRange.layerCount = 1;
-
-    vi.image = occ_atlas.image;
-    vi.format = VK_FORMAT_R8_UINT;
-    VK_CHECK(vkCreateImageView(device, &vi, nullptr,
-                               occ_view.init(device)));
-    VK_CHECK(vkCreateImageView(device, &vi, nullptr,
-                               occ_storage_view.init(device)));
-
-    vi.image = mat_atlas.image;
-    vi.format = VK_FORMAT_R8_UINT;
-    VK_CHECK(vkCreateImageView(device, &vi, nullptr,
-                               mat_view.init(device)));
-    VK_CHECK(vkCreateImageView(device, &vi, nullptr,
-                               mat_storage_view.init(device)));
-
-    vi.image = index_img.image;
-    vi.format = VK_FORMAT_R32_UINT;
-    VK_CHECK(vkCreateImageView(device, &vi, nullptr,
-                               index_view.init(device)));
-    VK_CHECK(vkCreateImageView(device, &vi, nullptr,
-                               index_storage_view.init(device)));
-
-    cpu_index.resize(static_cast<size_t>(map_dim.x) * map_dim.y * map_dim.z,
-                     0xFFFFFFFFu);
-    free_list.reserve(max_bricks);
-    lru.resize(max_bricks);
-    for (uint32_t i = 0; i < max_bricks; ++i)
-      free_list.push_back(max_bricks - 1 - i);
-  }
-
-  void destroy() {
-    index_storage_view.reset();
-    index_view.reset();
-    mat_storage_view.reset();
-    mat_view.reset();
-    occ_storage_view.reset();
-    occ_view.reset();
-    destroy_image3d(allocator, index_img);
-    destroy_image3d(allocator, mat_atlas);
-    destroy_image3d(allocator, occ_atlas);
-  }
-
-  // Acquire a brick at the given coordinate.  If necessary the least recently
-  // used brick is recycled.
-  uint32_t acquire(const BrickKey &key, uint32_t frame) {
-    auto it = bricks.find(key);
-    if (it != bricks.end()) {
-      lru[it->second.index].last_used = frame;
-      return it->second.index;
-    }
-
-    uint32_t slot = 0;
-    if (!free_list.empty()) {
-      slot = free_list.back();
-      free_list.pop_back();
-    } else {
-      // recycle least recently used brick
-      uint32_t lru_idx = 0;
-      uint32_t lru_frame = lru[0].last_used;
-      for (uint32_t i = 1; i < max_bricks; ++i) {
-        if (lru[i].last_used < lru_frame) {
-          lru_frame = lru[i].last_used;
-          lru_idx = i;
-        }
-      }
-      slot = lru_idx;
-      bricks.erase(lru[lru_idx].key);
-    }
-    Brick b{};
-    b.key = key;
-    b.index = slot;
-    b.last_used = frame;
-    lru[slot] = b;
-    bricks[key] = b;
-
-    // Update CPU lookup table
-    size_t offset =
-        (static_cast<size_t>(key.z) * map_dim.y + key.y) * map_dim.x + key.x;
-    if (offset < cpu_index.size()) cpu_index[offset] = slot;
-    return slot;
-  }
-
-  // Remove bricks that are further than 'radius' bricks from the camera.
-  void stream(const glm::vec3 &cam_pos, uint32_t frame, int radius) {
-    BrickKey cam_key{static_cast<int>(std::floor(cam_pos.x / brick_size)),
-                     static_cast<int>(std::floor(cam_pos.y / brick_size)),
-                     static_cast<int>(std::floor(cam_pos.z / brick_size))};
-
-    for (auto it = bricks.begin(); it != bricks.end();) {
-      const BrickKey &bk = it->first;
-      int dx = std::abs(bk.x - cam_key.x);
-      int dy = std::abs(bk.y - cam_key.y);
-      int dz = std::abs(bk.z - cam_key.z);
-      if (dx > radius || dy > radius || dz > radius) {
-        free_list.push_back(it->second.index);
-        size_t offset =
-            (static_cast<size_t>(bk.z) * map_dim.y + bk.y) * map_dim.x + bk.x;
-        if (offset < cpu_index.size()) cpu_index[offset] = 0xFFFFFFFFu;
-        it = bricks.erase(it);
-      } else {
-        lru[it->second.index].last_used = frame;
-        ++it;
-      }
-    }
-  }
-
-  VkImage occ_image() const { return occ_atlas.image; }
-  VkImage mat_image() const { return mat_atlas.image; }
-  VkImage index_image() const { return index_img.image; }
-};
-
-struct BrickParams {
-  glm::ivec3 brickCoord{0};
-  int brickIndex = 0;
-};
-
-struct DrawCtx {
-  RayPipeline *ray_pipe = nullptr;
-  VkDescriptorSet ray_dset = VK_NULL_HANDLE;
-  PresentPipeline *light_pipe = nullptr;
-  VkDescriptorSet light_dset = VK_NULL_HANDLE;
-  VkPipeline comp_pipe = VK_NULL_HANDLE;
-  VkPipelineLayout comp_layout = VK_NULL_HANDLE;
-  VkDescriptorSet comp_set = VK_NULL_HANDLE;
-  VkPipeline comp_l1_pipe = VK_NULL_HANDLE;
-  VkPipelineLayout comp_l1_layout = VK_NULL_HANDLE;
-  std::vector<VkDescriptorSet> comp_l1_sets;
-  std::vector<BrickParams> brick_params; // per-brick push constants
-  VkImage occ_image = VK_NULL_HANDLE;
-  VkImage mat_image = VK_NULL_HANDLE;
-  VkImage occ_l1_image = VK_NULL_HANDLE;
-  VkImage occ_l2_image = VK_NULL_HANDLE;
-  VkImage brick_ptr_image = VK_NULL_HANDLE;
-  VkImage g_albedo = VK_NULL_HANDLE;
-  VkImage g_normal = VK_NULL_HANDLE;
-  VkImage g_depth = VK_NULL_HANDLE;
-  VkImageView g_albedo_view = VK_NULL_HANDLE;
-  VkImageView g_normal_view = VK_NULL_HANDLE;
-  VkImageView g_depth_view = VK_NULL_HANDLE;
-  VkImage steps_image = VK_NULL_HANDLE;
-  VkImageView steps_view = VK_NULL_HANDLE;
-  VkBuffer steps_buffer = VK_NULL_HANDLE;
-  VkImageLayout occ_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkImageLayout mat_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkImageLayout occ_l1_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkImageLayout occ_l2_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkImageLayout g_albedo_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkImageLayout g_normal_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkImageLayout g_depth_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkImageLayout steps_layout = VK_IMAGE_LAYOUT_UNDEFINED;
-  VkExtent2D steps_dim{0, 0};
-  VkExtent2D ray_extent{0, 0};
-  VkExtent3D occ_dim{0, 0, 0};
-  VkExtent3D dispatch_dim{0, 0, 0};
-  VkExtent3D occ_l1_dim{0, 0, 0};
-  VkExtent3D dispatch_l1_dim{0, 0, 0};
-  VkExtent3D occ_l2_dim{0, 0, 0};
-  VkExtent3D dispatch_l2_dim{0, 0, 0};
-  VkImageView occ_view = VK_NULL_HANDLE;
-  VkImageView occ_l1_view = VK_NULL_HANDLE;
-  VkImageView occ_l1_storage_view = VK_NULL_HANDLE;
-  VkImageView occ_l2_view = VK_NULL_HANDLE;
-  VkImageView occ_l2_storage_view = VK_NULL_HANDLE;
-  uint32_t occ_levels = 0;
-  bool first_frame = true;
-};
-
-struct CameraUBO {
-  glm::mat4 inv_view_proj{1.0f};
-  glm::vec2 render_resolution{0.0f};
-  glm::vec2 output_resolution{0.0f};
-  float time = 0.0f;
-  float debug_normals = 0.0f;
-  float debug_level = 0.0f;
-  float debug_steps = 0.0f;
-  glm::vec4 pad{0.0f};
-};
-
-struct VoxelAABB {
-  glm::vec3 min{0.0f};
-  float pad0 = 0.0f;
-  glm::vec3 max{0.0f};
-  float pad1 = 0.0f;
-  glm::ivec3 dim{0};
-  int pad2 = 0;
-  glm::ivec3 occL1Dim{0};
-  int pad3 = 0;
-  glm::vec3 occL1CellSize{0.0f};
-  float pad4 = 0.0f;
-  glm::ivec3 occL2Dim{0};
-  int pad5 = 0;
-  glm::vec3 occL2CellSize{0.0f};
-  float pad6 = 0.0f;
-};
-
-struct VoxParams {
-  glm::ivec3 dim{0};
-  int frame = 0;
-  glm::vec3 volMin{0.0f};
-  float pad0 = 0.0f;
-  glm::vec3 volMax{0.0f};
-  float pad1 = 0.0f;
-  glm::vec3 boxCenter{0.0f};
-  float pad2 = 0.0f;
-  glm::vec3 boxHalf{0.0f};
-  float pad3 = 0.0f;
-  glm::vec3 sphereCenter{0.0f};
-  float sphereRadius = 0.0f;
-  int mode = 0;
-  int op = 0;
-  int noiseSeed = 0;
-  int material = 0;
-  float terrainFreq = 0.0f;
-  float grassDensity = 0.0f;
-  float treeDensity = 0.0f;
-  float flowerDensity = 0.0f;
-};
-
-struct BuildOccParams {
-  glm::ivec3 srcDim{0};
-  glm::ivec3 dstDim{0};
-  int blockSize = 1;
-};
-
-static VkShaderModule load_module(VkDevice dev, const std::string &path) {
-  std::ifstream f(path, std::ios::ate | std::ios::binary);
-  if (!f) {
-    spdlog::error("[vk] Failed to open SPIR-V: {}", path);
-    std::abort();
-  }
-  size_t size = static_cast<size_t>(f.tellg());
-  std::vector<char> data(size);
-  f.seekg(0);
-  f.read(data.data(), size);
-  VkShaderModuleCreateInfo ci{VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO};
-  ci.codeSize = data.size();
-  ci.pCode = reinterpret_cast<const uint32_t *>(data.data());
-  VkShaderModule mod = VK_NULL_HANDLE;
-  VK_CHECK(vkCreateShaderModule(dev, &ci, nullptr, &mod));
-  return mod;
-}
-
-static void record_present(VkCommandBuffer cmd, VkImage, VkImageView view,
-                           VkFormat, VkExtent2D extent, void *user) {
-  auto *ctx = static_cast<DrawCtx *>(user);
-  auto begin_label = [&](const char* name, const float color[4]) {
-    if (!cfg::kGpuMarkers || !vkCmdBeginDebugUtilsLabelEXT) return;
-    VkDebugUtilsLabelEXT l{VK_STRUCTURE_TYPE_DEBUG_UTILS_LABEL_EXT};
-    l.pLabelName = name;
-    for (int i = 0; i < 4; ++i) l.color[i] = color[i];
-    vkCmdBeginDebugUtilsLabelEXT(cmd, &l);
-  };
-  auto end_label = [&]() {
-    if (cfg::kGpuMarkers && vkCmdEndDebugUtilsLabelEXT)
-      vkCmdEndDebugUtilsLabelEXT(cmd);
-  };
-
-  begin_label("Voxel generation", kColorCompute);
-  VkImageMemoryBarrier2 pre[3]{};
-  for (int i = 0; i < 3; i++) {
-    pre[i].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-    pre[i].srcStageMask =
-        ctx->first_frame ? VK_PIPELINE_STAGE_2_NONE : VK_PIPELINE_STAGE_2_FRAGMENT_SHADER_BIT;
-    pre[i].srcAccessMask =
-        ctx->first_frame ? VK_ACCESS_2_NONE : VK_ACCESS_2_SHADER_SAMPLED_READ_BIT;
-    pre[i].dstStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-    pre[i].dstAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-    pre[i].newLayout = VK_IMAGE_LAYOUT_GENERAL;
-    pre[i].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    pre[i].subresourceRange.levelCount = 1;
-    pre[i].subresourceRange.layerCount = 1;
-  }
-  pre[0].image = ctx->occ_image;
-  pre[0].oldLayout = ctx->occ_layout;
-  pre[1].image = ctx->mat_image;
-  pre[1].oldLayout = ctx->mat_layout;
-  pre[2].image = ctx->occ_l1_image;
-  pre[2].oldLayout = ctx->occ_l1_layout;
-
-  VkDependencyInfo dep{VK_STRUCTURE_TYPE_DEPENDENCY_INFO};
-  dep.imageMemoryBarrierCount = 3;
-  dep.pImageMemoryBarriers = pre;
-  vkCmdPipelineBarrier2(cmd, &dep);
-
-  ctx->occ_layout = VK_IMAGE_LAYOUT_GENERAL;
-  ctx->mat_layout = VK_IMAGE_LAYOUT_GENERAL;
-  ctx->occ_l1_layout = VK_IMAGE_LAYOUT_GENERAL;
-
-  vkCmdBindPipeline(cmd, VK_PIPELINE_BIND_POINT_COMPUTE, ctx->comp_pipe);
-  vkCmdBindDescriptorSets(cmd, VK_PIPELINE_BIND_POINT_COMPUTE, ctx->comp_layout,
-                          0, 1, &ctx->comp_set, 0, nullptr);
-  uint32_t gx = (ctx->dispatch_dim.width + 7) / 8;
-  uint32_t gy = (ctx->dispatch_dim.height + 7) / 8;
-  uint32_t gz = (ctx->dispatch_dim.depth + 7) / 8;
-  for (const auto &bp : ctx->brick_params) {
-    vkCmdPushConstants(cmd, ctx->comp_layout, VK_SHADER_STAGE_COMPUTE_BIT, 0,
-                       sizeof(BrickParams), &bp);
-    vkCmdDispatch(cmd, gx, gy, gz);
-  }
-  end_label();
-
-  begin_label("Build occupancy", kColorCompute);
-  // Transition L0 outputs for sampling
-  VkImageMemoryBarrier2 mid[2]{};
-  for (int i = 0; i < 2; i++) {
-    mid[i].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-    mid[i].srcStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-    mid[i].srcAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-    mid[i].dstStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-    mid[i].dstAccessMask = VK_ACCESS_2_SHADER_SAMPLED_READ_BIT;
-    mid[i].newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    mid[i].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    mid[i].subresourceRange.levelCount = 1;
-    mid[i].subresourceRange.layerCount = 1;
-  }
-  mid[0].image = ctx->occ_image;
-  mid[0].oldLayout = ctx->occ_layout;
-  mid[1].image = ctx->mat_image;
-  mid[1].oldLayout = ctx->mat_layout;
-  dep.imageMemoryBarrierCount = 2;
-  dep.pImageMemoryBarriers = mid;
-  vkCmdPipelineBarrier2(cmd, &dep);
-
-  ctx->occ_layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-  ctx->mat_layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-
-  VkImage occ_images[3] = {ctx->occ_image, ctx->occ_l1_image, ctx->occ_l2_image};
-  VkExtent3D occ_dims[3] = {ctx->occ_dim, ctx->occ_l1_dim, ctx->occ_l2_dim};
-  VkImageLayout* occ_layouts[3] = {&ctx->occ_layout, &ctx->occ_l1_layout, &ctx->occ_l2_layout};
-
-  for (uint32_t level = 0; level < ctx->occ_levels - 1; ++level) {
-    VkImageMemoryBarrier2 b[2]{};
-    for (int i = 0; i < 2; ++i) {
-      b[i].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-      b[i].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-      b[i].subresourceRange.levelCount = 1;
-      b[i].subresourceRange.layerCount = 1;
-    }
-    b[0].srcStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-    b[0].srcAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-    b[0].dstStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-    b[0].dstAccessMask = VK_ACCESS_2_SHADER_SAMPLED_READ_BIT;
-    b[0].oldLayout = *occ_layouts[level];
-    b[0].newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    b[0].image = occ_images[level];
-
-    b[1].srcStageMask = ctx->first_frame ? VK_PIPELINE_STAGE_2_NONE : VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-    b[1].srcAccessMask = ctx->first_frame ? VK_ACCESS_2_NONE : VK_ACCESS_2_SHADER_WRITE_BIT;
-    b[1].dstStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-    b[1].dstAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-    b[1].oldLayout = *occ_layouts[level + 1];
-    b[1].newLayout = VK_IMAGE_LAYOUT_GENERAL;
-    b[1].image = occ_images[level + 1];
-
-    dep.imageMemoryBarrierCount = 2;
-    dep.pImageMemoryBarriers = b;
-    vkCmdPipelineBarrier2(cmd, &dep);
-
-    *occ_layouts[level] = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    *occ_layouts[level + 1] = VK_IMAGE_LAYOUT_GENERAL;
-
-    vkCmdBindPipeline(cmd, VK_PIPELINE_BIND_POINT_COMPUTE, ctx->comp_l1_pipe);
-    vkCmdBindDescriptorSets(cmd, VK_PIPELINE_BIND_POINT_COMPUTE,
-                            ctx->comp_l1_layout, 0, 1,
-                            &ctx->comp_l1_sets[level], 0, nullptr);
-
-    BuildOccParams params{};
-    params.srcDim = {static_cast<int>(occ_dims[level].width),
-                     static_cast<int>(occ_dims[level].height),
-                     static_cast<int>(occ_dims[level].depth)};
-    params.dstDim = {static_cast<int>(occ_dims[level + 1].width),
-                     static_cast<int>(occ_dims[level + 1].height),
-                     static_cast<int>(occ_dims[level + 1].depth)};
-    params.blockSize = static_cast<int>(occ_dims[level].width /
-                                        occ_dims[level + 1].width);
-    vkCmdPushConstants(cmd, ctx->comp_l1_layout, VK_SHADER_STAGE_COMPUTE_BIT,
-                       0, sizeof(params), &params);
-
-    gx = (occ_dims[level + 1].width + 7) / 8;
-    gy = (occ_dims[level + 1].height + 7) / 8;
-    gz = (occ_dims[level + 1].depth + 7) / 8;
-    vkCmdDispatch(cmd, gx, gy, gz);
-  }
-
-  end_label();
-
-  begin_label("Geometry", kColorGraphics);
-
-  VkImageMemoryBarrier2 geom[6]{};
-
-  // occ_l1_image: compute write -> fragment sampled read
-  geom[0].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-  geom[0].srcStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-  geom[0].srcAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-  geom[0].dstStageMask = VK_PIPELINE_STAGE_2_FRAGMENT_SHADER_BIT;
-  geom[0].dstAccessMask = VK_ACCESS_2_SHADER_SAMPLED_READ_BIT;
-  geom[0].oldLayout = ctx->occ_l1_layout;
-  geom[0].newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-  geom[0].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-  geom[0].subresourceRange.levelCount = 1;
-  geom[0].subresourceRange.layerCount = 1;
-  geom[0].image = ctx->occ_l1_image;
-
-  // occ_l2_image: compute write -> fragment sampled read
-  geom[1].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-  geom[1].srcStageMask = VK_PIPELINE_STAGE_2_COMPUTE_SHADER_BIT;
-  geom[1].srcAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-  geom[1].dstStageMask = VK_PIPELINE_STAGE_2_FRAGMENT_SHADER_BIT;
-  geom[1].dstAccessMask = VK_ACCESS_2_SHADER_SAMPLED_READ_BIT;
-  geom[1].oldLayout = ctx->occ_l2_layout;
-  geom[1].newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-  geom[1].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-  geom[1].subresourceRange.levelCount = 1;
-  geom[1].subresourceRange.layerCount = 1;
-  geom[1].image = ctx->occ_l2_image;
-
-  // steps image for fragment writes
-  geom[2].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-  geom[2].srcStageMask = ctx->first_frame ? VK_PIPELINE_STAGE_2_NONE : VK_PIPELINE_STAGE_2_TRANSFER_BIT;
-  geom[2].srcAccessMask = ctx->first_frame ? VK_ACCESS_2_NONE : VK_ACCESS_2_TRANSFER_WRITE_BIT;
-  geom[2].dstStageMask = VK_PIPELINE_STAGE_2_FRAGMENT_SHADER_BIT;
-  geom[2].dstAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-  geom[2].oldLayout = ctx->steps_layout;
-  geom[2].newLayout = VK_IMAGE_LAYOUT_GENERAL;
-  geom[2].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-  geom[2].subresourceRange.levelCount = 1;
-  geom[2].subresourceRange.layerCount = 1;
-  geom[2].image = ctx->steps_image;
-
-  // G-buffer images for color attachment writes
-  VkImageLayout* g_layouts[3] = {&ctx->g_albedo_layout, &ctx->g_normal_layout,
-                                 &ctx->g_depth_layout};
-  VkImage g_images[3] = {ctx->g_albedo, ctx->g_normal, ctx->g_depth};
-  for (int i = 0; i < 3; i++) {
-    int idx = i + 3;
-    geom[idx].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-    geom[idx].srcStageMask = ctx->first_frame ? VK_PIPELINE_STAGE_2_NONE : VK_PIPELINE_STAGE_2_FRAGMENT_SHADER_BIT;
-    geom[idx].srcAccessMask = ctx->first_frame ? VK_ACCESS_2_NONE : VK_ACCESS_2_SHADER_SAMPLED_READ_BIT;
-    geom[idx].dstStageMask = VK_PIPELINE_STAGE_2_COLOR_ATTACHMENT_OUTPUT_BIT;
-    geom[idx].dstAccessMask = VK_ACCESS_2_COLOR_ATTACHMENT_WRITE_BIT;
-    geom[idx].oldLayout = *g_layouts[i];
-    geom[idx].newLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    geom[idx].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    geom[idx].subresourceRange.levelCount = 1;
-    geom[idx].subresourceRange.layerCount = 1;
-    geom[idx].image = g_images[i];
-  }
-
-  dep.imageMemoryBarrierCount = 6;
-  dep.pImageMemoryBarriers = geom;
-  vkCmdPipelineBarrier2(cmd, &dep);
-
-  ctx->occ_l1_layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-  ctx->occ_l2_layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-  ctx->steps_layout = VK_IMAGE_LAYOUT_GENERAL;
-  ctx->g_albedo_layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-  ctx->g_normal_layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-  ctx->g_depth_layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-
-  // Geometry pass writing G-buffer
-  VkRenderingAttachmentInfo gAtt[3]{
-      {VK_STRUCTURE_TYPE_RENDERING_ATTACHMENT_INFO},
-      {VK_STRUCTURE_TYPE_RENDERING_ATTACHMENT_INFO},
-      {VK_STRUCTURE_TYPE_RENDERING_ATTACHMENT_INFO}};
-  VkClearValue gclear{};
-  gclear.color = {{0, 0, 0, 0}};
-  gAtt[0].imageView = ctx->g_albedo_view;
-  gAtt[0].imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-  gAtt[0].loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-  gAtt[0].storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-  gAtt[0].clearValue = gclear;
-  gAtt[1].imageView = ctx->g_normal_view;
-  gAtt[1].imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-  gAtt[1].loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-  gAtt[1].storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-  gAtt[1].clearValue = gclear;
-  gAtt[2].imageView = ctx->g_depth_view;
-  gAtt[2].imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-  gAtt[2].loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-  gAtt[2].storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-  gAtt[2].clearValue = gclear;
-
-  VkRenderingInfo gi{VK_STRUCTURE_TYPE_RENDERING_INFO};
-  gi.renderArea.offset = {0, 0};
-  gi.renderArea.extent = ctx->ray_extent;
-  gi.layerCount = 1;
-  gi.colorAttachmentCount = 3;
-  gi.pColorAttachments = gAtt;
-
-  vkCmdBeginRendering(cmd, &gi);
-
-  // Geometry pass renders the voxel scene at reduced resolution (ray_extent)
-  VkViewport vp_lo{};
-  vp_lo.x = 0.0f;
-  vp_lo.y = static_cast<float>(ctx->ray_extent.height);
-  vp_lo.width = static_cast<float>(ctx->ray_extent.width);
-  vp_lo.height = -static_cast<float>(ctx->ray_extent.height);
-  vp_lo.minDepth = 0.0f;
-  vp_lo.maxDepth = 1.0f;
-  vkCmdSetViewport(cmd, 0, 1, &vp_lo);
-  VkRect2D sc_lo{{0, 0}, ctx->ray_extent};
-  vkCmdSetScissor(cmd, 0, 1, &sc_lo);
-  vkCmdBindPipeline(cmd, VK_PIPELINE_BIND_POINT_GRAPHICS,
-                    ctx->ray_pipe->pipeline());
-  vkCmdBindDescriptorSets(cmd, VK_PIPELINE_BIND_POINT_GRAPHICS,
-                          ctx->ray_pipe->layout(), 0, 1, &ctx->ray_dset, 0,
-                          nullptr);
-  vkCmdDraw(cmd, 3, 1, 0, 0);
-  vkCmdEndRendering(cmd);
-  end_label();
-
-  begin_label("Readback", kColorTransfer);
-  VkImageSubresourceRange stepsRange{};
-  stepsRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-  stepsRange.levelCount = 1;
-  stepsRange.layerCount = 1;
-
-  VkImageMemoryBarrier2 steps_to_copy{};
-  steps_to_copy.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-  steps_to_copy.srcStageMask = VK_PIPELINE_STAGE_2_FRAGMENT_SHADER_BIT;
-  steps_to_copy.srcAccessMask = VK_ACCESS_2_SHADER_WRITE_BIT;
-  steps_to_copy.dstStageMask = VK_PIPELINE_STAGE_2_COPY_BIT;
-  steps_to_copy.dstAccessMask = VK_ACCESS_2_TRANSFER_READ_BIT;
-  steps_to_copy.oldLayout = ctx->steps_layout;
-  steps_to_copy.newLayout = VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL;
-  steps_to_copy.subresourceRange = stepsRange;
-  steps_to_copy.image = ctx->steps_image;
-
-  VkImageMemoryBarrier2 gpost[3]{};
-  VkImageLayout* g_layouts2[3] = {&ctx->g_albedo_layout, &ctx->g_normal_layout,
-                                  &ctx->g_depth_layout};
-  VkImage g_images2[3] = {ctx->g_albedo, ctx->g_normal, ctx->g_depth};
-  for (int i = 0; i < 3; i++) {
-    gpost[i].sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-    gpost[i].srcStageMask = VK_PIPELINE_STAGE_2_COLOR_ATTACHMENT_OUTPUT_BIT;
-    gpost[i].srcAccessMask = VK_ACCESS_2_COLOR_ATTACHMENT_WRITE_BIT;
-    gpost[i].dstStageMask = VK_PIPELINE_STAGE_2_FRAGMENT_SHADER_BIT;
-    gpost[i].dstAccessMask = VK_ACCESS_2_SHADER_SAMPLED_READ_BIT;
-    gpost[i].oldLayout = *g_layouts2[i];
-    gpost[i].newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    gpost[i].subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    gpost[i].subresourceRange.levelCount = 1;
-    gpost[i].subresourceRange.layerCount = 1;
-    gpost[i].image = g_images2[i];
-  }
-
-  VkImageMemoryBarrier2 readback_barriers[4] = {steps_to_copy, gpost[0], gpost[1], gpost[2]};
-  dep.imageMemoryBarrierCount = 4;
-  dep.pImageMemoryBarriers = readback_barriers;
-  vkCmdPipelineBarrier2(cmd, &dep);
-
-  ctx->steps_layout = VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL;
-  for (int i = 0; i < 3; ++i)
-    *g_layouts2[i] = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-
-  VkBufferImageCopy bic{};
-  bic.imageSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-  bic.imageSubresource.layerCount = 1;
-  bic.imageExtent = {ctx->steps_dim.width, ctx->steps_dim.height, 1};
-  vkCmdCopyImageToBuffer(cmd, ctx->steps_image,
-                         VK_IMAGE_LAYOUT_TRANSFER_SRC_OPTIMAL,
-                         ctx->steps_buffer, 1, &bic);
-
-  VkImageMemoryBarrier2 steps_to_clear{};
-  steps_to_clear.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER_2;
-  steps_to_clear.srcStageMask = VK_PIPELINE_STAGE_2_COPY_BIT;
-  steps_to_clear.srcAccessMask = VK_ACCESS_2_TRANSFER_READ_BIT;
-  steps_to_clear.dstStageMask = VK_PIPELINE_STAGE_2_CLEAR_BIT;
-  steps_to_clear.dstAccessMask = VK_ACCESS_2_TRANSFER_WRITE_BIT;
-  steps_to_clear.oldLayout = ctx->steps_layout;
-  steps_to_clear.newLayout = VK_IMAGE_LAYOUT_GENERAL;
-  steps_to_clear.subresourceRange = stepsRange;
-  steps_to_clear.image = ctx->steps_image;
-  dep.imageMemoryBarrierCount = 1;
-  dep.pImageMemoryBarriers = &steps_to_clear;
-  vkCmdPipelineBarrier2(cmd, &dep);
-
-  ctx->steps_layout = VK_IMAGE_LAYOUT_GENERAL;
-
-  VkClearColorValue zero{{0, 0, 0, 0}};
-  vkCmdClearColorImage(cmd, ctx->steps_image, VK_IMAGE_LAYOUT_GENERAL, &zero, 1,
-                       &stepsRange);
-  end_label();
-
-  begin_label("Postprocess", kColorGraphics);
-
-  // Lighting pass to final image
-  VkRenderingAttachmentInfo color{VK_STRUCTURE_TYPE_RENDERING_ATTACHMENT_INFO};
-  color.imageView = view;
-  color.imageLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-  color.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-  color.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-  VkClearValue clear;
-  clear.color = {{0.06f, 0.07f, 0.10f, 1.0f}};
-  color.clearValue = clear;
-
-  VkRenderingInfo ri{VK_STRUCTURE_TYPE_RENDERING_INFO};
-  ri.renderArea.offset = {0, 0};
-  ri.renderArea.extent = extent;
-  ri.layerCount = 1;
-  ri.colorAttachmentCount = 1;
-  ri.pColorAttachments = &color;
-
-  // Final lighting/present pass upscales to the swapchain extent
-  vkCmdBeginRendering(cmd, &ri);
-
-  VkViewport vp_hi{};
-  vp_hi.x = 0.0f;
-  vp_hi.y = static_cast<float>(extent.height);
-  vp_hi.width = static_cast<float>(extent.width);
-  vp_hi.height = -static_cast<float>(extent.height);
-  vp_hi.minDepth = 0.0f;
-  vp_hi.maxDepth = 1.0f;
-  vkCmdSetViewport(cmd, 0, 1, &vp_hi);
-
-  VkRect2D sc_hi{{0, 0}, extent};
-  vkCmdSetScissor(cmd, 0, 1, &sc_hi);
-
-  vkCmdBindPipeline(cmd, VK_PIPELINE_BIND_POINT_GRAPHICS,
-                    ctx->light_pipe->pipeline());
-  vkCmdBindDescriptorSets(cmd, VK_PIPELINE_BIND_POINT_GRAPHICS,
-                          ctx->light_pipe->layout(), 0, 1, &ctx->light_dset, 0,
-                          nullptr);
-  vkCmdDraw(cmd, 3, 1, 0, 0);
-  vkCmdEndRendering(cmd);
-  end_label();
-
-  ctx->first_frame = false;
-}
-
 int run() {
+
   init_logging();
   log_boot_banner("engine");
